DKC x Mario (SNES): fix cutscene/ending swap
At the end of the game, there's a lot of changes in values at the address that usually seems to hold the "Mario died" sprite. This sometimes hits the value that triggers a swap. Force the swap condition to check that the in-level timer is running. This counter may be useful elsewhere as well.