Vice: Project Doom (NES): back to hp/lives
Because iframes sometimes start on the frame after HP loss (possibly due to lag?), they're not ideal. Fortunately, there is a reliable indicator for the current stage after all, that works with or without the level select cheat. Revert to using the usual hp/lives function with a proper exception to shuffle only on deaths during shooting galleries.