Dick Tracy (NES): prefer infinite health
Dick Tracy (NES) does not use lives, and continues are punishing because of having to collect clues again. You can refill first aid kits, but they don't automatically deploy if you lose your last HP. Smight recommends just restoring health on swap. Note, just using 2 for HP and for first aid kits is not workable, as shooting unarmed people deducts two health, and you could be damaged quickly enough despite grace to game over.

Noted that #24 could give options for what to restore on swapping in and that #14 and #18 are also relevant for choosing how to implement this.